refactor(card_image_display): split CardImageDisplay into mixin package
Move widgets/panels/card_image_display.py into a package with a shared
frame.py that builds the UI plus properties/handlers mixins for the
display's getters, image-navigation callbacks, and fade-transition
worker.
