+++ widgets/panels/card_image_display/__init__.py
@@ -1,0 +1,7 @@
+"""Card image display widget package."""
+
+from __future__ import annotations
+
+from widgets.panels.card_image_display.frame import CardImageDisplay
+
+__all__ = ["CardImageDisplay"]
